LPD-55085 Avoiding parameter leaks
This fixes the issue for this specific case. However the issue is
broader as the proxy being created here:

https://github.com/liferay/liferay-portal/blob/772256c9238e157fa7525e33ca6dba79c91fe8c7/modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/template/RESTClientTemplateContextContributor.java#L98-L103

Is using the original request as default object thus delegating every
other method to it unless overriden in the delegate. We should probably
rethink that a little bit... because what we really need from the
original request should be just the authentication information.